FIX: - Bỏ chức năng nhập mã tra cứu - Bổ sung thêm thông tin tồn kho trong chức năng tra cứu và chức năng mua sắm vật tư phần danh sách vật tư và chi tiết vật tư

+++ ESPlatform.QRCode.IMS.Domain/Entities/QlvtVatTuTonKho.cs
@@ -20,7 +20,7 @@ public partial class QlvtVatTuTonKho
     /// <summary>
     /// Số lượng tồn
     /// </summary>
-    public int? OnhandQuantity { get; set; }
+    public decimal? OnhandQuantity { get; set; }
 
     /// <summary>
     /// Mã kho phụ
